Fix size of chat_id field in DB

+++ database/migrations/2025_02_17_134805_create_pipelines_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('pipelines', function (Blueprint $table) {
             $table->id();
-            $table->integer('chat_id')->unique();
+            $table->bigInteger('chat_id')->unique();
             $table->string('command');
             $table->integer('stage');
             $table->timestamps();
